restructing tests and directory cleanup
